fix(charts): constrain the value, not the pixel, in NumericScale.trackValue
`trackValue(v, offset, constrain)` clamped the raw pixel coordinate `v`
instead of the computed `value` when `constrain` was true, mismatching
TimeScale.trackValue. With a pixel range wider than the value range, an
in-range value whose pixel fell outside [min, max] was wrongly clamped.

- Extract the private NumericScale calculator into its own React-free
  module so it can be unit-tested in isolation (public API unchanged; it
  is not re-exported from the axis barrel).
- Add a regression test covering constrained/unconstrained trackValue.
